bio-mark-duplicates: make LibraryBagSize report bag size instead of number of library duplicates
Summary:
The "LibraryBagSize" name is misleading because the value is equal to
the number of PCR duplicates, whereas a bag contains the duplicates
and the primary.

This change adds one to the LibraryBagSize (LD tag) to include a count for the primary read.

Test Plan: Updated and ran existing unit tests.

Reviewers: jrosenfield, cchang

Reviewed By: jrosenfield, cchang

Subscribers: escott

Maniphest Tasks: T30406

Differential Revision: https://phabricator.grailbio.com/D40154

fbshipit-source-id: 1d730ce