bio-mark-duplicates: implement fine-grained high-coverage regions with subsampling
Summary:
bio-mark-duplicates: implement fine-grained high-coverage regions with subsampling

The previous version drops a readpair if the read or mate maps to a
shard that contains a high-coverage region. This strategy drops more
reads than necessary; it unnecessarily drops reads that map to
low-coverage regions if those low-coverage regions happen to be in the
same shard as a high-coverage region. This is worse when shards cover
large regions of the genome (like when using PAM files).

This change:
* reduces the number of read pairs that are dropped to only those that map to a high-coverage region.
* changes to subsample instead of drop all reads that are aligned to a high-coverage region.
* if R1 and R2 both overlap different high-coverage region, subsample the pair according to the region of higher coverage.
* makes it possible to disable dropping any high-coverage reads.
* changes the high-coverage region output file to use 1-based genomic coordinates.

Test Plan:
Modified some test cases, added some new ones.
I tested with several of the CCGA samples and R&D samples that have high coverage areas, and the high-coverage areas that are identified, and also subsampling looks to be in the right ballpark.

Reviewers: akokate

Reviewed By: akokate

Subscribers: escott

Maniphest Tasks: T30530

Differential Revision: https://phabricator.grailbio.com/D43415

fbshipit-source-id: ead0fbf